+++ Loop/Views/CarbEntryView.swift
@@ -80,8 +80,13 @@ struct CarbEntryView: View, HorizontalSizeClassOverride {
                 }
                 
                 if viewModel.selectedFavoriteFoodLastEaten != nil, FeatureFlags.allowExperimentalFeatures {
-                    favoriteFoodInsightsCard
-                        .padding(.top, 8)
+                    FavoriteFoodInsightsCardView(
+                        showFavoriteFoodInsights: $showFavoriteFoodInsights,
+                        foodName: viewModel.selectedFavoriteFood?.name,
+                        lastEatenDate: viewModel.selectedFavoriteFoodLastEaten,
+                        relativeDateFormatter: viewModel.relativeDateFormatter
+                    )
+                    .padding(.top, 8)
                 }
                 
                 let isBolusViewActive = Binding(get: { viewModel.bolusViewModel != nil }, set: { _, _ in viewModel.bolusViewModel = nil })
@@ -242,7 +247,7 @@ extension CarbEntryView {
                 .padding(.horizontal, 26)
             
             VStack(spacing: 10) {
-                if !viewModel.favoriteFoods.isEmpty, isNewEntry {
+                if !viewModel.favoriteFoods.isEmpty {
                     VStack {
                         HStack {
                             Text("Choose Favorite:")
@@ -321,57 +326,6 @@ extension CarbEntryView {
         clearExpandedRow()
         self.showAddFavoriteFood = false
         viewModel.onFavoriteFoodSave(food)
-    }
-}
-
-// MARK: - Favorite Food Insights Card
-extension CarbEntryView {
-    private var favoriteFoodInsightsCard: some View {
-        Button(action: {
-            showFavoriteFoodInsights = true
-        }) {
-            VStack(spacing: 10) {
-                HStack(spacing: 4) {
-                    Image(systemName: "sparkles")
-
-                    Text("Favorite Food Insights")
-                }
-                .font(.headline)
-                .foregroundColor(.accentColor)
-                .frame(maxWidth: .infinity, alignment: .leading)
-                
-                if let foodName = viewModel.selectedFavoriteFood?.name,
-                   let lastEatenDate = viewModel.selectedFavoriteFoodLastEaten {
-                    let relativeTime = viewModel.relativeDateFormatter.localizedString(for: lastEatenDate, relativeTo: Date())
-                    let attributedFoodDescription = attributedFoodInsightsDescription(for: foodName, timeAgo: relativeTime)
-                    
-                    Text(attributedFoodDescription)
-                        .foregroundColor(.primary)
-                        .multilineTextAlignment(.center)
-                }
-            }
-            .padding(.vertical, 12)
-            .padding(.horizontal)
-            .background(CardBackground())
-            .overlay {
-                RoundedRectangle(cornerRadius: 10, style: .continuous)
-                    .strokeBorder(Color.accentColor, lineWidth: 2)
-            }
-            .padding(.horizontal)
-            .contentShape(Rectangle())
-        }
-    }
-    
-    private func attributedFoodInsightsDescription(for food: String, timeAgo: String) -> AttributedString {
-        var attributedString = AttributedString("You last ate ")
-        
-        var foodString = AttributedString(food)
-        foodString.inlinePresentationIntent = .stronglyEmphasized
-        
-        attributedString.append(foodString)
-        attributedString.append(AttributedString(" \(timeAgo)\n Tap to see more"))
-        
-        return attributedString
     }
 }
